Refactor image field to named function
Apparently django is unable to serialise lambda expressions, which I
think was causing migrations issues. Not sure how it was working after
the last commit (because I tested it...) but I ran into some other
issues and had to re-run all my migrations, and then after that it was
refusing to work with the lambda expression, so here we are.