Add watchlist url pathway
Logged in users can now visit the 'Watchlist' url end-point to
see a simple list of listings that they have added to their watchlist.
Each of the listings is a link, that redirects to the listing's page.

The watchlist page is implemented almost identically to the index page,
however since the purpose is significantly different, and will likely
have many additional features when it comes to adding styling and final
features that I kept it separate to avoid complications.