Set up auction app logging
Using a custom configuration, logs can now be created for the auctions
app. By default INFO and above will be logged to the console, while
DEBUG and above will be saved to a timed rotating file handler. This
allows important monitoring information to written to the console, but
if a deeper inspection is required the log files can be checked.

Logging level for console logger can also be set by the 'LOGLEVEL'
environment variable.