Add top categories selection and display
All views will now display the top three categories at the top of the
layout. I had to manually add the query and variable to each view as I
couldn't figure out how to add it via a wrapper function. There's
probably a way, but I didn't see it being worth the investment of time.