Add misc category
In order to add a 'breadcrumbs' widget to the listing detail page I need
to ensure that every listing is linked to a category. By adding a misc
category, which is the default, I don't have to put a load of if else
checks for empty categories, so this makes things nice and clean.