Refactor create bid path with error handling
Redesigned how the create bid pathway works, with it no longer being a
url of its own, but being a form that can be submitted with the listing
page. This way I can render error messages for incorrect submissions.

I designed the forms to have an action attribute which will determine
which function needs to be run. This is easily extendible to the other
pathways because of some mapping dictionaries.